Renamed gdsc-test-api-generator to gdsc-test-generator
