--- gdsc-test-api-generator/README.md
@@ -1,9 +1,0 @@
-GDSC Test API Generator
-=======================
-
-Generates API classes for the GDSC Test project.
-
-[![License: GPL v3](https://img.shields.io/badge/License-GPL%20v3-blue.svg)](https://www.gnu.org/licenses/gpl-3.0)
-
-This generator uses [String Template](http://www.stringtemplate.org/) and
-[Picoli](https://picocli.info/).
