--- gdsc-test-api-generator/src/main/java/uk/ac/sussex/gdsc/test/api/generator/Triple.java
@@ -1,57 +1,0 @@
-/*-
- * #%L
- * Genome Damage and Stability Centre Test Utilities
- *
- * Generates API classes for the GDSC Test project.
- * %%
- * Copyright (C) 2018 Alex Herbert
- * %%
- * This program is free software: you can redistribute it and/or modify
- * it under the terms of the GNU General Public License as
- * published by the Free Software Foundation, either version 3 of the
- * License, or (at your option) any later version.
- *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU General Public License for more details.
- *
- * You should have received a copy of the GNU General Public
- * License along with this program.  If not, see
- * <http://www.gnu.org/licenses/gpl-3.0.html>.
- * #L%
- */
-
-package uk.ac.sussex.gdsc.test.api.generator;
-
-/**
- * A generic triple.
- *
- * @param <A> the generic type of the first value
- * @param <B> the generic type of the second value
- * @param <C> the generic type of the third value
- */
-public class Triple<A, B, C> {
-
-  /** The first value. */
-  public final A first;
-
-  /** The second value. */
-  public final B second;
-
-  /** The third value. */
-  public final C third;
-
-  /**
-   * Instantiates a new triple.
-   *
-   * @param first the first value
-   * @param second the second value
-   * @param third the third value
-   */
-  public Triple(A first, B second, C third) {
-    this.first = first;
-    this.second = second;
-    this.third = third;
-  }
-}
